requirements.txt
matplotlib==3.3.4
Pillow==7.0.0
numpy==1.20.1
torch==1.8.0
torchvision==0.9.0
higher==0.2.1

